+++ test/tests/tiny_comm_test.cpp
@@ -98,9 +98,9 @@ TEST_GROUP(tiny_comm)
   static constexpr uint16_t calculate_crc(uint16_t seed, uint8_t byte)
   {
     uint16_t crc = seed;
-    byte = crc >> 8 ^ byte;
-    byte ^= byte >> 4;
-    return (crc << 8) ^ ((uint16_t)(byte << 12)) ^ ((uint16_t)(byte << 5)) ^ ((uint16_t)byte);
+    byte = (uint8_t)(crc >> 8 ^ byte);
+    byte ^= (uint8_t)(byte >> 4);
+    return (uint16_t)((crc << 8) ^ ((uint16_t)(byte << 12)) ^ ((uint16_t)(byte << 5)) ^ ((uint16_t)byte));
   }
 
 #define crc_msb(...) _crc_msb<__VA_ARGS__>()
